+++ packages/nimble-components/src/rich-text/models/tests/markdown-serializer.spec.ts
@@ -337,12 +337,13 @@ Plain text 3`);
 <user:1> `);
         });
 
-        // Intermittent, see https://github.com/ni/nimble/issues/2219
-        xit('Mention node', async () => {
+        it('Mention node', async () => {
             await appendUserMentionConfiguration(element, [
                 { key: 'user:1', displayName: 'username1' }
             ]);
+
             await commitFirstMentionBoxOptionIntoEditor('@');
+
             expect(element.getMarkdown()).toEqual('<user:1> ');
         });
 
@@ -358,7 +359,8 @@ Plain text 3`);
             expect(element.getMarkdown()).toEqual('<user:1> <user:2> ');
         });
 
-        it('Multiple Mention node of different type', async () => {
+        // Intermittent on Webkit (at least), see https://github.com/ni/nimble/issues/2426
+        it('Multiple Mention node of different type #SkipWebkit', async () => {
             await appendUserMentionConfiguration(element, [
                 { key: 'user:1', displayName: 'username1' }
             ]);